refactored profile screen
